+++ tests/acceptance/conftest.py
@@ -7,11 +7,11 @@ from upadup.updater import UpadupUpdater
 
 @pytest.fixture
 def update_from_text(tmp_path):
-    def _updatefunc(content):
+    def _updatefunc(content: str, freeze: bool = False):
         config_path = tmp_path / ".pre-commit-config.yaml"
         config_path.write_text(textwrap.dedent(content))
 
-        updater = UpadupUpdater(path=config_path)
+        updater = UpadupUpdater(path=config_path, freeze=freeze)
         updater.run()
         updater.apply_updates()